+++ packages/extension/src/view/devtools/components/tabContent/cookies/cookiesListing/cookieDetails/index.tsx
