+++ packages/extension/src/view/devtools/components/tabContent/cookies/cookiesListing/cookieDetails/stories/cookieDetails.stories.ts
