+++ packages/extension/src/view/devtools/components/tabContent/cookies/cookiesListing/cookieTable/index.tsx
